backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
chromadb==0.4.22
sentence-transformers==2.7.0
groq==0.4.1
python-dotenv==1.0.0
huggingface-hub>=0.20.0
